plugin/templates.vim: Lots of bugfixes to get things to work again.
After the cleanup a lot of small things needed to be fixed to get
everything to work again.  The attribute-parsing code now also checks for
duplicated attributes and deals with a dictionary directly instead of first
putting everything in a list and then turning it into a dictionary.  It’s
so-so of a gain, but at least it works well.  Also fixed the exception
handling to not intercept pure scripting-errors that Vim tries to report.