plugin/templates.vim: Added license-file placeholder
The license-file placeholder has been added.  It works really well, but it
would be better if placeholders were instantiated once per invocation with
the correct position in the template and so on, so that error-reporting and
the strpart() call in the expand() method of LicensePlaceholder could done
a bit more reliably.

Also added a default-value parameter to s:borgval().

A bug in how templates were replaced was fixed as well.  Now we use
append() in combination to setline(), as setline() will replace lines if
given more than one line.  Which isn’t really very strange, just not what
we want.